refac: extract Hasher and Digest into separate files
