+++ crates/common/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod digest;
 pub mod hashchain;
+pub mod hasher;
 pub mod keys;
 pub mod operation;
 pub mod tree;
