+++ crates/storage/src/database.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use auto_impl::auto_impl;
 use jmt::storage::{TreeReader, TreeWriter};
-use prism_common::tree::Digest;
+use prism_common::digest::Digest;
 use prism_errors::{DatabaseError, PrismError};
 
 #[auto_impl(&, Box, Arc)]
