+++ crates/storage/src/inmemory.rs
@@ -3,7 +3,7 @@ use jmt::{
     storage::{LeafNode, Node, NodeBatch, NodeKey, TreeReader, TreeWriter},
     KeyHash, OwnedValue, Version,
 };
-use prism_common::tree::Digest;
+use prism_common::digest::Digest;
 use prism_errors::DatabaseError;
 use std::{
     collections::HashMap,
